branching on type addded to Spectra
